+++ src/main/java/org/kuali/hr/lm/timeoff/service/SystemScheduledTimeOffService.java
@@ -49,7 +49,7 @@ public interface SystemScheduledTimeOffService {
 	 * @param sstoHours
 	 * @return
 	 */
-	public BigDecimal calculateSysSchTimeOffHours(Job job, Long sstoHours);
+    public BigDecimal calculateSysSchTimeOffHours(Job job, BigDecimal sstoHours);
 
     List<SystemScheduledTimeOff> getSystemScheduledTimeOffs(Date fromEffdt, Date toEffdt, String earnCode, String fromAccruedDate,
                                                             String toAccruedDate, String fromSchTimeOffDate, String toSchTimeOffDate, String active, String showHist);
